pyproject.toml
[tool.poetry]
name = "tcbot"
version = "0.1.0"
description = ""
authors = ["dff9th <dff9th@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
tweepy = {git = "https://github.com/tweepy/tweepy.git"}
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
jedi = "^0.18.0"
pylint = "^2.6.0"
black = "^20.8b1"
pytest = "^6.2.2"

[tool.poetry.scripts]
tcbot = "tcbot.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
